notify_debouncer_full

Function new_debouncer

pub fn new_debouncer<F: DebounceEventHandler>(
    timeout: Duration,
    tick_rate: Option<Duration>,
    event_handler: F,
) -> Result<Debouncer<RecommendedWatcher, RecommendedCache>, Error>

Short function to create a new debounced watcher with the recommended debouncer and the built-in file ID cache.

Timeout is the amount of time after which a debounced event is emitted.

If tick_rate is None, notify will select a tick rate that is 1/4 of the provided timeout.